Add commands for items and item_taxes

+++ src-tauri/src/schema.rs
@@ -43,7 +43,7 @@ table! {
     taxes (id) {
         id -> Text,
         name -> Text,
-        rate -> BigInt,
+        rate -> Integer,
         description -> Nullable<Text>,
         created_at -> Timestamp,
         updated_at -> Timestamp,
